move decodeHtmlEntities to stringUtils
