--- test/unit/specs/services/decodeHtmlEntities.spec.js
@@ -1,12 +1,0 @@
-import decodeHtmlEntities from 'services/decodeHtmlEntities'
-
-describe('services/decodeHtmlEntities', () => {
-  let pairs = [
-    ['&amp;', '&']
-  ]
-  pairs.forEach(([from, to]) => {
-    it(`converts "${from}" -> "${to}"`, () => {
-      expect(decodeHtmlEntities(from)).to.equal(to)
-    })
-  })
-})
