+++ test/unit/specs/services/stringUtils.spec.js
@@ -1,4 +1,4 @@
-import { escape, autolink, nl2br } from 'services/stringUtils'
+import { escape, autolink, nl2br, decodeHtmlEntities } from 'services/stringUtils'
 
 describe('services/stringUtils', () => {
   it('escape', () => {
@@ -12,5 +12,16 @@ describe('services/stringUtils', () => {
 
   it('nl2br', () => {
     expect(nl2br('hello\nworld\n')).to.equal('hello<br>world<br>')
+  })
+
+  context('decodeHtmlEntities', () => {
+    let pairs = [
+      ['&amp;', '&']
+    ]
+    pairs.forEach(([from, to]) => {
+      it(`converts "${from}" -> "${to}"`, () => {
+        expect(decodeHtmlEntities(from)).to.equal(to)
+      })
+    })
   })
 })
